Separate op types, wrap opdef utils in Operators
