+++ src/index.ts
@@ -1,3 +1,4 @@
 
-export { Expr, Result, ExprCtx, registerOp } from './core/expressions'
+export { Expr, Result, ExprCtx } from './core/expressions'
+export { Operators, Op, OpDef } from './core/operators'
 export { parseExpr, parseExprList } from './core/parse'
